+++ schemas/config.json
@@ -25,10 +25,8 @@
     },
     "parser_install_directories": {
       "description": "A list of strings representing directories to search for parsers, of the form `<lang>.(so|dll|dylib)` or `tree-sitter-<lang>.wasm`.\n\nSupports environment variable expansion of the form `${VAR}`.",
-      "type": [
-        "array",
-        "null"
-      ],
+      "default": [],
+      "type": "array",
       "items": {
         "type": "string"
       }
