+++ src/test_helpers.rs
@@ -91,7 +91,7 @@ pub mod helpers {
             HashMap::new()
         };
         let options = Arc::new(tokio::sync::RwLock::new(Options {
-            parser_install_directories: None,
+            parser_install_directories: Default::default(),
             parser_aliases: None,
             language_retrieval_patterns: None,
             valid_captures,
